% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autobahn-package.R
\docType{package}
\name{autobahn-package}
\alias{autobahn}
\alias{autobahn-package}
\title{autobahn: Interface to Autobahn App API}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

An R client for Autobahn App API, \url{https://autobahn.api.bund.dev/}.
}
\author{
\strong{Maintainer}: Markus Döring \email{m4rkus.doering@gmail.com} [copyright holder]

}
\keyword{internal}
