% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autobahn.R
\name{autobahn_electric_charging_station}
\alias{autobahn_electric_charging_station}
\title{Get list of electric charging stations for a given autobahn}
\usage{
autobahn_electric_charging_station(
  roadId = NULL,
  return = c("as-is", "simplified")
)
}
\arguments{
\item{roadId}{A character string. Identifier of a autobahn.}

\item{return}{Should function return result as is, or simplify the
response? Defaults to \code{as-is}.}
}
\value{
A list.
}
\description{
Get list of electric charging stations for a given autobahn
}
\examples{
A5 <- autobahn_electric_charging_station("A5")
A5 <- autobahn_electric_charging_station("A5", return = "simplified")
}
