% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autobahn.R
\name{autobahn_parking_lorry}
\alias{autobahn_parking_lorry}
\title{Get list of parking lorries for a given autobahn}
\usage{
autobahn_parking_lorry(roadId = NULL)
}
\arguments{
\item{roadId}{A character vector. Identifier of a autobahn.}
}
\value{
A list.
}
\description{
Get list of parking lorries for a given autobahn
}
\examples{
A1 <- autobahn_parking_lorry("A1")
}
