% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autobahn.R
\name{autobahn_roadworks}
\alias{autobahn_roadworks}
\title{Get list of construction sites for a given autobahn}
\usage{
autobahn_roadworks(roadId = NULL)
}
\arguments{
\item{roadId}{A character vector. Identifier of a autobahn.}
}
\value{
A list.
}
\description{
Get list of construction sites for a given autobahn
}
\examples{
A1 <- autobahn_roadworks("A1")
}
