% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autobahn.R
\name{autobahn_warning}
\alias{autobahn_warning}
\title{Get list of warnings for a given autobahn}
\usage{
autobahn_warning(roadId = NULL)
}
\arguments{
\item{roadId}{A character vector. Identifier of a autobahn.}
}
\value{
A list.
}
\description{
Get list of warnings for a given autobahn
}
\examples{
A1 <- autobahn_warning("A1")
}
