% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autobahn.R
\name{autobahnen}
\alias{autobahnen}
\title{Get list of all autobahns in Germany}
\usage{
autobahnen()
}
\value{
A character vector.
}
\description{
Get list of all autobahns in Germany
}
\details{
The function calls
\url{https://verkehr.autobahn.de/o/autobahn/}
and returns a vector of all autobahns in Germany, see:
\url{https://autobahn.api.bund.dev/}.

More information on Germany's autobahns can be found on
\href{https://en.wikipedia.org/wiki/List_of_autobahns_in_Germany}{wikipedia.org}
}
\examples{
autobahnen()

}
