% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autobahn.R
\name{autobahn_closure}
\alias{autobahn_closure}
\title{Get list of closures for a given autobahn}
\usage{
autobahn_closure(roadId = NULL)
}
\arguments{
\item{roadId}{A character vector. Identifier of a autobahn.}
}
\value{
A list.
}
\description{
Get list of closures for a given autobahn
}
\examples{
A1 <- autobahn_closure("A1")
}
