% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autobahn.R
\name{autobahn_warning}
\alias{autobahn_warning}
\title{Get list of warnings for a given autobahn}
\usage{
autobahn_warning(roadId = NULL, return = c("as-is", "simplified"))
}
\arguments{
\item{roadId}{A character string. Identifier of a autobahn.}

\item{return}{Should function return result as is, or simplify the
response? Defaults to \code{as-is}.}
}
\value{
A list.
}
\description{
Get list of warnings for a given autobahn
}
\examples{
A3 <- autobahn_warning("A3")
A3 <- autobahn_warning("A3", return = "simplified")
}
