% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autobahn.R
\name{autobahnen}
\alias{autobahnen}
\title{Get list of all autobahns in Germany}
\usage{
autobahnen()
}
\value{
A character vector.
}
\description{
Get list of all autobahns in Germany
}
\details{
The function returns a vector of all autbahns in Germany. Under the hood it
calls \url{https://verkehr.autobahn.de/o/autobahn/}.
For more information on the endpoint see:
\url{https://autobahn.api.bund.dev/}.

More information on Germany's autobahns can be found here:
\href{https://en.wikipedia.org/wiki/List_of_autobahns_in_Germany}{en.wikipedia.org/wiki/List_of_autobahns_in_Germany}
}
\examples{
autobahnen()

}
